+++ routes/api/expenses.php
@@ -4,14 +4,14 @@ use App\Http\Controllers\Api\ExpenseController;
 use Illuminate\Support\Facades\Route;
 
 Route::middleware('auth:api')->prefix('expenses')->group(function () {
-    Route::get('/statistics',            [ExpenseController::class, 'statistics']);
-    Route::get('/',                      [ExpenseController::class, 'index']);
-    Route::post('/',                     [ExpenseController::class, 'store']);
-    Route::get('/{id}',                  [ExpenseController::class, 'show']);
-    Route::put('/{id}',                  [ExpenseController::class, 'update']);
-    Route::delete('/{id}',               [ExpenseController::class, 'destroy']);
-    Route::post('/{id}/documents',       [ExpenseController::class, 'uploadDocuments']);
-    Route::post('/{id}/receipts',        [ExpenseController::class, 'uploadReceipts']);
-    Route::get('/{id}/media',            [ExpenseController::class, 'getMedia']);
-    Route::delete('/{id}/media/{mediaId}',[ExpenseController::class, 'deleteMedia']);
+    Route::get('/statistics',            [ExpenseController::class, 'statistics'])->middleware('permission:view-expense');
+    Route::get('/',                      [ExpenseController::class, 'index'])->middleware('permission:view-expense');
+    Route::post('/',                     [ExpenseController::class, 'store'])->middleware('permission:create-expense');
+    Route::get('/{id}',                  [ExpenseController::class, 'show'])->middleware('permission:view-expense');
+    Route::put('/{id}',                  [ExpenseController::class, 'update'])->middleware('permission:edit-expense');
+    Route::delete('/{id}',               [ExpenseController::class, 'destroy'])->middleware('permission:delete-expense');
+    Route::post('/{id}/documents',       [ExpenseController::class, 'uploadDocuments'])->middleware('permission:edit-expense');
+    Route::post('/{id}/receipts',        [ExpenseController::class, 'uploadReceipts'])->middleware('permission:edit-expense');
+    Route::get('/{id}/media',            [ExpenseController::class, 'getMedia'])->middleware('permission:view-expense');
+    Route::delete('/{id}/media/{mediaId}',[ExpenseController::class, 'deleteMedia'])->middleware('permission:edit-expense');
 });
